refactor: organize components about `dayGridWeek`
